Move biggerapp to just example_app folder
