+++ example_app/templates/index.html
