+++ example_app/templates/layout.html
