+++ example_app/templates/pagination.html
